+++ src/components/Home.css
@@ -1,6 +1,6 @@
 /* Without setting height the slides will take up the height of the slide's content */
 ion-slides {
-  height: 100%;
+  height: auto;
 }
 
 .parrafoSlide{
@@ -43,5 +43,10 @@ p b {
 .card{
   background-color: #F8F7F2;
 }
-
-
+.title{
+  font-size: 14px;
+  font-weight: bold;
+}
+.carText{
+  justify-content: flex-start;
+}
